[MODORDERS-535] - Fix error with checkin cache
