+++ src/test/java/org/folio/ApiTestSuite.java
@@ -15,6 +15,7 @@ import org.folio.orders.events.handlers.ReceiptStatusConsistencyTest;
 import org.folio.orders.events.handlers.ReceiveOrderStatusChangeHandlerTest;
 import org.folio.orders.utils.HelperUtilsTest;
 import org.folio.orders.utils.validators.CompositePoLineValidationUtilTest;
+import org.folio.orders.utils.validators.LocationsAndPiecesConsistencyValidatorTest;
 import org.folio.orders.utils.validators.OngoingOrderValidatorTest;
 import org.folio.rest.core.RestClientTest;
 import org.folio.rest.impl.CheckinReceivingApiTest;
@@ -270,6 +271,11 @@ public class ApiTestSuite {
 
   @Nested
   class CompositePoLineValidationUtilNested extends CompositePoLineValidationUtilTest {
+
+  }
+
+  @Nested
+  class LocationsAndPiecesConsistencyValidatorTestNested extends LocationsAndPiecesConsistencyValidatorTest {
 
   }
 }
